machine-learning/requirements.txt
flask
python-dotenv
transformers
pillow
sentence-transformers
chromadb
groq
mysql-connector-python